test(e2e): move update handler tests to the mandated tests/ layout
`.agents/testing.md`'s STRICT RULE forbids inline `#[cfg(test)] mod`
blocks in `crates/*/src`, and `build-support/check-inline-tests.sh`
enforces it as the first step of `make check-daemon`. My local
verification ran the cargo commands directly and skipped that gate, so
the violation only surfaced in CI.

`handlers/update.rs` becomes `handlers/update/mod.rs` with the tests in
`handlers/update/tests/systemd.rs`. No behaviour change.
