+++ source/daemon/crates/wardnet-test-agent/src/server/handlers/update/mod.rs
@@ -414,33 +414,4 @@ fn show_property(stdout: &str, key: &str) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn show_property_extracts_a_present_key() {
-        let out = "ActiveState=failed\nResult=exit-code\nExecMainStatus=1\n";
-        assert_eq!(show_property(out, "ActiveState"), "failed");
-        assert_eq!(show_property(out, "Result"), "exit-code");
-        assert_eq!(show_property(out, "ExecMainStatus"), "1");
-    }
-
-    #[test]
-    fn show_property_is_empty_for_a_missing_key() {
-        assert_eq!(show_property("ActiveState=active\n", "Result"), "");
-    }
-
-    #[test]
-    fn show_property_does_not_match_a_key_that_is_only_a_suffix() {
-        // `ExecMainStatus` must not be satisfied by `MainStatus=`; the
-        // prefix match is anchored to the start of the line.
-        assert_eq!(show_property("MainStatus=3\n", "ExecMainStatus"), "");
-    }
-
-    #[test]
-    fn allowlist_covers_only_the_two_units_the_suite_drives() {
-        assert!(ALLOWED_UNITS.contains(&"wardnetd.service"));
-        assert!(ALLOWED_UNITS.contains(&"wardnet-postupgrade.service"));
-        assert!(!ALLOWED_UNITS.contains(&"wardnet-test-agent.service"));
-    }
-}
+mod tests;
